requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --strip-extras
#
aniso8601==9.0.1
    # via flask-restful
blinker==1.9.0
    # via flask
click==8.1.7
    # via flask
colorama==0.4.6
    # via click
deprecated==1.2.14
    # via pikepdf
flask==3.0.3
    # via
    #   -r requirements.in
    #   flask-restful
flask-restful==0.3.10
    # via -r requirements.in
img2pdf==0.4.4
    # via -r requirements.in
itsdangerous==2.2.0
    # via flask
jinja2==3.1.4
    # via flask
lxml==5.3.0
    # via pikepdf
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
packaging==24.2
    # via pikepdf
pikepdf==9.4.0
    # via img2pdf
pillow==11.0.0
    # via
    #   img2pdf
    #   pikepdf
python-decouple==3.8
    # via -r requirements.in
pytz==2024.2
    # via flask-restful
six==1.16.0
    # via flask-restful
werkzeug==3.1.3
    # via flask
wrapt==1.16.0
    # via deprecated
